+++ src/Database/PostgreSQL/PQTypes/Internal/Exception.hs
@@ -14,6 +14,8 @@ import Database.PostgreSQL.PQTypes.SQL.Class
 data DBException = forall e sql. (E.Exception e, Show sql) => DBException
   { dbeQueryContext :: !sql
   -- ^ Last SQL query that was executed.
+  , dbeBackendPid :: !Int
+  -- ^ Process ID of the server process attached to the current session.
   , dbeError :: !e
   -- ^ Specific error.
   , dbeCallStack :: CallStack
@@ -24,11 +26,12 @@ deriving instance Show DBException
 instance E.Exception DBException
 
 -- | Rethrow supplied exception enriched with given SQL.
-rethrowWithContext :: (HasCallStack, IsSQL sql) => sql -> E.SomeException -> IO a
-rethrowWithContext sql (E.SomeException e) =
+rethrowWithContext :: (HasCallStack, IsSQL sql) => sql -> Int -> E.SomeException -> IO a
+rethrowWithContext sql pid (E.SomeException e) =
   E.throwIO
     DBException
       { dbeQueryContext = sql
+      , dbeBackendPid = pid
       , dbeError = e
       , dbeCallStack = callStack
       }