+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # hpqtypes-1.12.0.0 (????-??-??)
 * Drop support for GHC 8.8.
 * Attach `CallStack` to `DBException`.
+* Store ID of the server process attached to the current session.
 
 # hpqtypes-1.11.1.2 (2023-11-08)
 * Support multihost setups and the `connect_timeout` parameter in the connection